+++ lib/mixins/block_movement_collision.dart
@@ -8,7 +8,9 @@ mixin BlockMovementCollision on Movement {
   final _collisionUtil = CollisionUtil();
 
   Rect? _shapeRectNormalized;
-  Direction? lasDirectionCollision;
+  Direction? _lasDirectionCollision;
+  Vector2 _lastDisplacementCollision = Vector2.zero();
+  Vector2 midPoint = Vector2.zero();
 
   bool onBlockMovement(
     Set<Vector2> intersectionPoints,
@@ -20,53 +22,52 @@ mixin BlockMovementCollision on Movement {
   void onBlockedMovement(
     PositionComponent other,
     Direction? direction,
-    Vector2 lastDisplacement,
   ) {
     final reverseDisplacement = _adjustDisplacement(
-      lastDisplacement,
+      _lastDisplacementCollision,
       direction,
     );
 
-    position += reverseDisplacement * -1;
+    superPosition = position - reverseDisplacement;
     angle = lastAngle;
+
     stopFromCollision(
-      isX: reverseDisplacement.x.abs() > 0,
-      isY: reverseDisplacement.y.abs() > 0,
+      isX: direction?.isSameXDirection(_lastDisplacementCollision.x) == true,
+      isY: direction?.isSameYDirection(_lastDisplacementCollision.y) == true,
     );
   }
-
-  Vector2 midPoint = Vector2.zero();
 
   @override
   void onCollision(Set<Vector2> intersectionPoints, PositionComponent other) {
     super.onCollision(intersectionPoints, other);
-    bool stopOtherMovement = true;
+    if (other is Sensor) return;
     bool stopMovement = other is GameComponent
         ? onBlockMovement(intersectionPoints, other)
         : true;
-    if (stopMovement && other is BlockMovementCollision) {
-      stopOtherMovement = other.onBlockMovement(intersectionPoints, this);
+    bool stopOtherMovement = other is BlockMovementCollision
+        ? other.onBlockMovement(intersectionPoints, this)
+        : true;
+    if (!stopMovement || !stopOtherMovement) {
+      return;
     }
-    if (stopMovement && stopOtherMovement && other is! Sensor) {
-      if (_shapeRectNormalized != null) {
-        midPoint = intersectionPoints.reduce(
-          (value, element) => value + element,
-        );
-        midPoint /= intersectionPoints.length.toDouble();
-        midPoint = midPoint - position;
-        midPoint.lerp(_shapeRectNormalized!.center.toVector2(), 0.2);
+    _lastDisplacementCollision = lastDisplacement.clone();
+    if (_shapeRectNormalized != null) {
+      midPoint = intersectionPoints.reduce(
+        (value, element) => value + element,
+      );
+      midPoint /= intersectionPoints.length.toDouble();
+      midPoint = midPoint - position;
+      midPoint.lerp(_shapeRectNormalized!.center.toVector2(), 0.2);
 
-        lasDirectionCollision = _collisionUtil.getDirectionCollision(
-          _shapeRectNormalized!,
-          midPoint,
-        );
+      _lasDirectionCollision = _collisionUtil.getDirectionCollision(
+        _shapeRectNormalized!,
+        midPoint,
+      );
 
-        onBlockedMovement(
-          other,
-          lasDirectionCollision,
-          lastDisplacement.clone(),
-        );
-      }
+      onBlockedMovement(
+        other,
+        _lasDirectionCollision,
+      );
     }
   }