+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.2
+- `BleedingPixel` improvements.
+- Adds `orientation` param in `getZoomFromMaxVisibleTile`.
+- Adds methods `showStroke` and `hideStroke` in `SimpleDirectionAnimation`
+- Adds methods `showSpriteStroke` and `hideSpriteStroke` in mixin `UseSprite`
+- Adds methods `showAnimationStroke` and `hideAnimationStroke` in mixin `UseSpriteAnimation`
+- BREAK: `BlockMovementCollision` improvements. Remove `lastDisplacement` from `onBlockedMovement` method.
+- BREAK: methos `seeAndMoveTo...` improvements. `notObserved` now return a bool (true to stop move).
+
 ## 3.0.1
 - Update Flame to `1.10.0`.
 - `Movement` improvements.